Add user-management component
