+++ src/app/components/Records/user-management/user-management.component.html
@@ -1,1 +1,130 @@
-<p>user-management works!</p>
+<h2 class="page-header">User Management</h2>
+<mat-card appearance="outlined">
+  <mat-card-header>
+    <div class="example-button-container">
+      <!-- mat-fab
+      extended -->
+      <button (click)="openDialog('Add')" class="btn btn-primary">
+        <mat-icon>add</mat-icon>
+        Add user
+      </button>
+    </div>
+  </mat-card-header>
+  <mat-card-content>
+    <table
+      mat-table
+      [dataSource]="dataSource"
+      class="mat-elevation-z8 rounded-table"
+    >
+      <!-- Select Column -->
+      <ng-container matColumnDef="select">
+        <th mat-header-cell *matHeaderCellDef>
+          <mat-checkbox
+            (change)="$event ? toggleAllRows() : null"
+            [checked]="selection.hasValue() && isAllSelected()"
+            [indeterminate]="selection.hasValue() && !isAllSelected()"
+          >
+          </mat-checkbox>
+        </th>
+        <td mat-cell *matCellDef="let row">
+          <mat-checkbox
+            (click)="$event.stopPropagation()"
+            (change)="$event ? selection.toggle(row) : null"
+            [checked]="selection.isSelected(row)"
+          >
+          </mat-checkbox>
+        </td>
+      </ng-container>
+
+      <!-- ID Column -->
+      <ng-container matColumnDef="id">
+        <th mat-header-cell *matHeaderCellDef>NO.</th>
+        <td mat-cell *matCellDef="let element">
+          {{ element.id | numberPadding : 4 }}
+        </td>
+      </ng-container>
+
+      <!-- Username Column -->
+      <ng-container matColumnDef="username">
+        <th mat-header-cell *matHeaderCellDef>USERNAME</th>
+        <td mat-cell *matCellDef="let element">
+          {{ element.username | titlecase }}
+        </td>
+      </ng-container>
+
+      <!-- Password Column -->
+      <ng-container matColumnDef="password">
+        <th mat-header-cell *matHeaderCellDef>PASSWORD</th>
+        <td mat-cell *matCellDef="let element">{{ element.password }}</td>
+      </ng-container>
+
+      <!-- Date Added Column -->
+      <ng-container matColumnDef="dateadded">
+        <th mat-header-cell *matHeaderCellDef>DATE ADDED</th>
+        <td mat-cell *matCellDef="let element">
+          {{ element.dateadded | date : "d MMM y" }}
+        </td>
+      </ng-container>
+
+      <!-- User Group Column -->
+      <ng-container matColumnDef="usergroup">
+        <th mat-header-cell *matHeaderCellDef>USER GROUP</th>
+        <td mat-cell *matCellDef="let element">
+          {{ element.usergroup | titlecase }}
+        </td>
+      </ng-container>
+
+      <!-- Status Column -->
+      <ng-container matColumnDef="status">
+        <th mat-header-cell *matHeaderCellDef>STATUS</th>
+        <td mat-cell *matCellDef="let element">
+          <div
+            class="badge"
+            [style.background]="
+              element.status == 'active'
+                ? 'rgba(28, 204, 62, 0.15)'
+                : 'rgba(116, 116, 128, 0.15)'
+            "
+            [style.color]="element.status == 'active' ? '#1ccc3e' : '#747480'"
+          >
+            {{ element.status | uppercase }}
+          </div>
+        </td>
+      </ng-container>
+
+      <!-- Action Column -->
+      <ng-container matColumnDef="action">
+        <th mat-header-cell *matHeaderCellDef>ACTION</th>
+        <td mat-cell *matCellDef="let element">
+          <div class="example-button-row">
+            <div class="example-flex-container">
+              <button
+                mat-icon-button
+                aria-label="Example icon button with a vertical three dot icon"
+                [matMenuTriggerFor]="menu"
+                #menuTrigger
+              >
+                <mat-icon>more_vert</mat-icon>
+              </button>
+              <mat-menu #menu="matMenu">
+                <button mat-menu-item (click)="openDialog('Edit')">Edit</button>
+                <button mat-menu-item (click)="showDeleteActionDialog()">
+                  Delete
+                </button>
+              </mat-menu>
+            </div>
+          </div>
+        </td>
+      </ng-container>
+
+      <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+      <tr mat-row *matRowDef="let row; columns: displayedColumns"></tr>
+    </table>
+    <mat-paginator
+      [pageSizeOptions]="[5, 10, 20]"
+      showFirstLastButtons
+      aria-label="Select page of periodic elements"
+    >
+    </mat-paginator>
+  </mat-card-content>
+</mat-card>
