--- src/app/components/user-management/user-management.component.css
@@ -1,53 +1,0 @@
-* {
-    color: white;
-    font-family: "Poppins", sans-serif;
-    font-weight: 400;
-    font-style: normal;
-    /* font-size: 0.75rem; */
-  }
-  
-  .page-header {
-    font-weight: 600;
-    font-size: 1rem;
-  }
-  
-  /* .btn-primary {
-    border-radius: 6px;
-    background-color: #0a84ff;
-    border: 1px solid #0a84ff;
-    color: white;
-    padding: 12px;
-    display: flex;
-    align-items: center;
-  } */
-  
-  mat-card-header {
-    border-bottom: 1px solid rgba(120, 120, 128, 0.36);
-    padding-bottom: 1rem;
-    margin-bottom: 1rem;
-  }
-  .mat-mdc-table {
-    border-radius: 6px;
-    background-color: transparent;
-    margin-bottom: 1rem;
-  }
-  
-  .mat-mdc-cell,
-  .mat-mdc-header-cell {
-    text-align: center;
-  }
-  .badge {
-    padding: 6px 8px;
-    text-align: center;
-    border-radius: 6px;
-    font-weight: 500;
-    /* font-size: 0.7rem; */
-  }
-  
-  .mat-mdc-paginator {
-    font-family: "Poppins", sans-serif;
-    font-weight: 400;
-    font-style: normal;
-    /* font-size: 0.75rem; */
-  }
-
