+++ src/app/pipes/number-padding.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { NumberPaddingPipe } from './number-padding.pipe';
+
+describe('NumberPaddingPipe', () => {
+  it('create an instance', () => {
+    const pipe = new NumberPaddingPipe();
+    expect(pipe).toBeTruthy();
+  });
+});
