+++ src/custom-theme.scss
@@ -12,7 +12,7 @@
 // Define the theme object.
 $CNCSportal-theme: mat.define-theme((
   color: (
-    theme-type: light,
+    theme-type: dark,
     primary: mat.$azure-palette,
     tertiary: mat.$blue-palette,
   ),
