+++ src/styles.css
@@ -1,4 +1,46 @@
 /* You can add global styles to this file, and also import other style files */
 
-html, body { height: 100%; }
-body { margin: 0; font-family: Roboto, "Helvetica Neue", sans-serif; background-color: #000000;}
+html,
+body {
+  height: 100%;
+}
+body {
+  margin: 0;
+  font-family: Roboto, "Helvetica Neue", sans-serif;
+  background-color: #000000;
+}
+
+html,
+body {
+  height: 100%;
+}
+body {
+  margin: 0;
+  font-family: Roboto, "Helvetica Neue", sans-serif;
+}
+
+.btn {
+  color: white;
+  border-radius: 6px;
+  padding: 12px;
+  display: flex;
+  align-items: center;
+}
+
+.btn-confirm {
+  background-color: #1ccc3e;
+}
+.btn-cancel {
+  background-color: #f0352d;
+}
+.btn-primary {
+  background-color: #0a84ff;
+}
+.btn-delete {
+  background-color: #ff9f0a;
+}
+
+mat-icon {
+  font-family: "Material Icons" !important;
+  font-size: 1.5rem;
+}
